requirements.txt
python-geoip==1.2
python-geoip-geolite2==2015.0303
setuptools==20.1.1
